fix(pg): Fixed error logs for multi node cron jobs.
When multiple nodes were participating as workers and a cron job was to
be enqueued, all nodes would attempt to enqueue it at the same time. As
a result all but one node would successfully enqueue the job and the
other nodes would log a message about a duplicate job. In any other
scenario this would be a log that would be good to surface, but it is
not a significant error message here as the job was still enqueued
properly.

Resolves #76