fix(pg): Adding tests for multiple workers and cron jobs.
This adds a test that proves that asserts that in a ~1 second period
and a cron that happens every second. The job will only be consumed once
by a worker pool of 8 workers.

This also resolves an issue where connections were not being released at
the end of tests that caused tests to outright fail in CI and timeout.